feat: Slash Commandのデバックコマンドの削除
