+++ src/main/kotlin/dev/m2en/citation/command/DebugCommand.kt
@@ -1,34 +1,21 @@
 package dev.m2en.citation.command
 
-import dev.kord.common.entity.Snowflake
-import dev.kord.core.behavior.interaction.response.DeferredEphemeralMessageInteractionResponseBehavior
 import dev.kord.core.behavior.interaction.response.respond
-import dev.kord.core.entity.interaction.GuildChatInputCommandInteraction
-import dev.kord.core.kordLogger
+import dev.kord.core.event.interaction.GuildMessageCommandInteractionCreateEvent
 import dev.kord.rest.builder.message.EmbedBuilder
 
-object DebugCommand: InteractionCommandInterface {
+suspend fun GuildMessageCommandInteractionCreateEvent.onDebugMessageCommand() {
+    val targetMessage = interaction.channel.getMessage(interaction.target.id)
 
-    override suspend fun onCommand(interaction: GuildChatInputCommandInteraction, responseBehavior: DeferredEphemeralMessageInteractionResponseBehavior) {
-        try {
-            val command = interaction.command
-            val recruit = interaction.channel.getMessage(Snowflake(command.strings["target"]!!))
+    interaction.deferEphemeralResponse().respond { embeds = MutableList(1) { buildEmbed(targetMessage.content) } }
+}
 
-            responseBehavior.respond { embeds = MutableList(1) { buildEmbed(recruit.content) } }
-            kordLogger.info("デバック: ${interaction.user.tag} が メッセージID:${command.strings["target-message"]} をデバックしました")
-        } catch (e: Exception) {
-            responseBehavior.respond { content = "> **エラー:** デバックに失敗しました" }
-            kordLogger.error("デバック: ${interaction.user.tag} がリクエストしたメッセージデバックに失敗しました (Reason: ${e.message})")
-        }
-    }
+private fun buildEmbed(targetContent: String): EmbedBuilder {
+    val debugRecruit = targetContent.replace("```", "'''")
+    val embed = EmbedBuilder()
 
-    private fun buildEmbed(recruit: String): EmbedBuilder {
-        val embed = EmbedBuilder()
-        embed.title = "メッセージデバック"
-        val debugRecruit = recruit.replace("```", "'''")
-        embed.description = "```\n${debugRecruit}\n```"
-        return embed
-    }
+    embed.title = "メッセージデバック"
+    embed.description = "```\n${debugRecruit}\n```"
 
-
+    return embed
 }
