--- src/main/kotlin/dev/m2en/citation/command/DebugMessageCommand.kt
@@ -1,21 +1,0 @@
-package dev.m2en.citation.command
-
-import dev.kord.core.behavior.interaction.response.respond
-import dev.kord.core.event.interaction.GuildMessageCommandInteractionCreateEvent
-import dev.kord.rest.builder.message.EmbedBuilder
-
-suspend fun GuildMessageCommandInteractionCreateEvent.onDebugMessageCommand() {
-    val targetMessage = interaction.channel.getMessage(interaction.target.id)
-
-    interaction.deferEphemeralResponse().respond { embeds = MutableList(1) { buildEmbed(targetMessage.content) } }
-}
-
-private fun buildEmbed(targetContent: String): EmbedBuilder {
-    val debugRecruit = targetContent.replace("```", "'''")
-    val embed = EmbedBuilder()
-
-    embed.title = "メッセージデバック"
-    embed.description = "```\n${debugRecruit}\n```"
-
-    return embed
-}
